simplify timeout tests and move them to session_test.go
